+++ rustls/src/lib.rs
@@ -272,6 +272,7 @@ mod record_layer;
 mod stream;
 #[cfg(feature = "tls12")]
 mod tls12;
+mod tls13;
 mod vecbuf;
 mod verify;
 #[cfg(test)]
